rename common in utils to renderer to better capture what this does
