+++ packages/client/src/utils/renderer/common.js
